Authentication with protected tweets done

+++ backend/controllers/tweetController.js
@@ -5,7 +5,7 @@ const Tweet = require('../models/tweetModel')
 // @route   GET /api/tweets
 // @access  Private
 const getTweets = asyncHandler(async (req, res) => {
-    const tweets = await Tweet.find()
+    const tweets = await Tweet.find({ user: req.user.id })
 
     res.status(200).json(tweets)
 })
@@ -20,6 +20,7 @@ const createTweet = asyncHandler(async (req, res) => {
     }
 
     const newTweet = await Tweet.create({
+        user: req.user.id,
         text: req.body.text
     })
 
@@ -30,12 +31,26 @@ const createTweet = asyncHandler(async (req, res) => {
 // @route   UPDATE /api/tweets/:id
 // @access  Private
 const updateTweet = asyncHandler(async (req, res) => {
+    if (!req.body.text) {
+        res.status(400)
+        throw new Error('Please add some text')
+    }
+
     const id = req.params.id
+    if (!req.user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
 
     const tweet = await Tweet.findById(id)
     if (!tweet) {
         res.status(400)
         throw new Error('Tweet not found!')
+    }
+
+    if (tweet.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Not Authorized')
     }
 
     const updatedTweet = await Tweet.findByIdAndUpdate(id, req.body, { new: true })
@@ -49,10 +64,20 @@ const updateTweet = asyncHandler(async (req, res) => {
 const deleteTweet = asyncHandler(async (req, res) => {
     const id = req.params.id
 
+    if (!req.user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
     const tweet = await Tweet.findById(id)
     if (!tweet) {
         res.status(400)
         throw new Error('Tweet not found!')
+    }
+
+    if (tweet.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Not Authorized')
     }
 
     await Tweet.findByIdAndDelete(id)
